refactor(template): use jsonRes for api responses
